Deleted unused DAGs
