+++ cornflow-dags/DAG/update_all_schemas.py
@@ -77,24 +77,67 @@ def get_all_schemas():
     return schemas_new
 
 
-def update_schemas(**kwargs):
-    schemas = get_all_schemas()
+def get_all_example_data():
+    apps = get_new_apps()
+    if len(apps):
+        print("Found the following apps: {}".format([app.name for app in apps]))
+    else:
+        print("No apps were found to update")
+    example_data_new = {}
+
+    for app in apps:
+        tests = app.test_cases
+        print(app, tests)
+        n = 1
+        example = dict()
+        for t in tests:
+            if isinstance(t, dict):
+                instance = f"instance_{n}"
+                example[instance] = t
+
+            elif isinstance(t, tuple):
+                instance = f"instance_{n}"
+                solution = f"solution_{n}"
+                example[instance] = t[0]
+                example[solution] = t[1]
+            n = n + 1
+
+        example_data_new[f"z_{app.name}_examples"] = example
+
+    print("Found the following new apps: {}".format([app.name for app in apps]))
+    return example_data_new
+
+
+def update_all_variables(**kwargs):
+    # first we delete all variables (this helps to keep it clean)
+    with create_session() as session:
+        current_vars = set(var.key for var in session.query(Variable))
+        for _var in current_vars:
+            Variable.delete(_var, session)
+
     # we update all schemas that we found:
+    schemas = get_all_schemas()
+
     for key, value in schemas.items():
+        Variable.set(key=key, value=value, serialize_json=True)
+
+    # we update all examples that we found:
+    example_data = get_all_example_data()
+    for key, value in example_data.items():
         Variable.set(key=key, value=value, serialize_json=True)
 
 
 dag = DAG(
-    "update_all_schemas", default_args=default_args, catchup=False, tags=["internal"]
+    "update_all_schemas", default_args=default_args, catchup=False, tags=["internal"], schedule_interval="@hourly"
 )
 
 update_schema2 = PythonOperator(
     task_id="update_all_schemas",
     provide_context=True,
-    python_callable=update_schemas,
+    python_callable=update_all_variables,
     dag=dag,
 )
 
 
 if __name__ == "__main__":
-    update_schemas()
+    update_all_variables()
